the basic version of GarFL

+++ TF-IDF/run.py
@@ -3,21 +3,17 @@ import os
 import pickle
 import shutil
 import sys
+from typing import List
+from unittest import result
 
 from langchain_community.retrievers import TFIDFRetriever
 from langchain_core.documents import Document
 
 from Evaluation.evaluate import evaluate_others
-from functions.d4j import (
-    check_out,
-    get_failed_tests,
-    get_properties,
-    parse_sbfl,
-    run_all_tests,
-)
+from functions.d4j import check_out, get_failed_tests, get_properties, run_all_tests
+from functions.sbfl import parse_sbfl
 from preprocess.index_builder import ProjectIndexBuilder
-from TestAnalysis.analyzer import TestAnalyzer
-from Utils.model import set_models
+from Query.query import QueryGenerator
 from Utils.path_manager import PathManager
 
 root = os.path.dirname(__file__)
@@ -25,14 +21,16 @@ sys.path.append(root)
 
 def main():
     parser = argparse.ArgumentParser(description='argparse')
-    parser.add_argument('--config', type=str, default="LMStudio+jina+openai",
+    parser.add_argument('--config', type=str, default="TF-IDF",
                         help="Name of config, which is used to load configuration under Config/")
     parser.add_argument('--version', type=str, default="2.0.1",
                         help="Version of defects4j")
     parser.add_argument('--project', type=str, default="Closure",
                         help="Name of project, your debug result will be generated in DebugResult/d4jversion_project_bugID")
-    parser.add_argument('--bugID', type=int, default=1,
+    parser.add_argument('--bugID', type=int, default=3,
                         help="Prompt of software")
+    parser.add_argument('--subproj', type=str, required=False, default="",
+                        help="The subproject of the project")
     parser.add_argument('--clear', type=bool, default=True,
                         help="If clear the checkout project")
     args = parser.parse_args()
@@ -66,55 +64,59 @@ def main():
     run_all_tests(path_manager, test_failure_obj)
 
     # ----------------------------------------
-    #          Test Analysis
+    #          Query Generation
     # ----------------------------------------
 
-    path_manager.logger.info("[Test Analysis] start...")
-    test_analyzer = TestAnalyzer(path_manager)
-    test_analyzer.analyze(test_failure_obj)
+    path_manager.logger.info("[Query Generation] start...")
+    query_generator = QueryGenerator(path_manager)
+    queries: List[str] = query_generator.generate(test_failure_obj)
 
     # ----------------------------------------
     #          SBFL results
     # ----------------------------------------
 
-    sbfl_res = parse_sbfl(path_manager)
+    sbfl_res = parse_sbfl(path_manager.sbfl_file)
 
     # ----------------------------------------
     #          Load Index
     # ----------------------------------------
 
-    path_manager.logger.info("[load data] start...")
-    index_builder = ProjectIndexBuilder(path_manager)
-    nodes = index_builder.build_nodes(sbfl_res)
-    documents = [Document(page_content=node.text) for node in nodes]
-    retriever = TFIDFRetriever.from_documents(documents, k=path_manager.retrieve_top_n)
-    
-    # ----------------------------------------
-    #          Retrieve
-    # ----------------------------------------
-    
-    path_manager.logger.info("[Retrieve] start...")
-    results = []
-    for test_class in test_failure_obj.test_classes:
-        for test_case in test_class.test_cases:
-            result = retriever.invoke(
-                "\n".join([
-                    test_case.test_output,
-                    test_case.stack_trace,
-                    test_case.test_method.text
-                ])
-            )
-            results.append(result)
+    if os.path.exists(path_manager.retrieved_nodes_file):
+        with open(path_manager.retrieved_nodes_file, 'rb') as f:
+            result_nodes = pickle.load(f)
+    else:
+        path_manager.logger.info("[load data] start...")
+        index_builder = ProjectIndexBuilder(path_manager)
+        nodes = index_builder.build_nodes([sbfl_res])
+        documents = [Document(page_content=node.text) for node in nodes]
+        if documents == []:
+            result_nodes = []
+        else:
+            retriever = TFIDFRetriever.from_documents(documents, k=path_manager.retrieve_top_n)
+            
+            # ----------------------------------------
+            #          Retrieve
+            # ----------------------------------------
+            
+            path_manager.logger.info("[Retrieve] start...")
+            all_queries = " ".join(queries)
+            result_nodes = retriever._get_relevant_documents(all_queries, run_manager=None)
+            with open(path_manager.retrieved_nodes_file, 'wb') as f:
+                pickle.dump(result_nodes, f)
+
 
     # ----------------------------------------
     #          Evaluate
     # ----------------------------------------
     
-    evaluate_others(path_manager, results, test_failure_obj)
+    evaluate_others(path_manager, result_nodes, test_failure_obj)
     
     if args.clear:
-        shutil.rmtree(path_manager.buggy_path)
-        shutil.rmtree(path_manager.fixed_path)
+        if os.path.exists(path_manager.buggy_path):
+            shutil.rmtree(path_manager.buggy_path)
+        if os.path.exists(path_manager.fixed_path):
+            shutil.rmtree(path_manager.fixed_path)
+
 
 if __name__ == "__main__":
     main()
